+++ src/main/java/com/greenblog/model/User.java
@@ -17,7 +17,7 @@ public class User {
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -31,7 +31,7 @@ public class User {
     public User() {
     }
 
-    public User(int id, String name, String lastName, String email) {
+    public User(Long id, String name, String lastName, String email) {
         this.id = id;
         this.name = name;
         this.lastName = lastName;
@@ -39,11 +39,11 @@ public class User {
     }
 
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
